Refactored wait executor
